Fix for issue #3411925: START inside an EXEC SQL is going into Water.
Updated the scoring file to remove this case for the START token.

Should really consider whether it still is meaningful to keep this scoring system alive... It was a useful way to help identify cases which a grammar rule missed out on. But the grammar has grown so much by now that maintaining these rules may no longer be worth the effort.

git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@93 07b9d407-957a-469a-9197-ae794bbef2a3