Integrated the BatchIt and ShowIt apps into one.
So BatchIt and ShowIt have been removed and replaced by one Koopa GUI app.

The new GUI uses tabs to show open files (the equivalent of ShowIt) and an overview (the equivalent of BatchIt).
Any Cobol source file will open in a new tab and will also be listed in the overview.
If you open a folder then it will just go in batch mode and show the results in the overview.
Double clicking items in the overview will also open up the file in a tab.


git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@117 07b9d407-957a-469a-9197-ae794bbef2a3