Search is now regex-based and case-insensitive.
Re. #5 FR for Show It: Add "search" and "go to line" .

As requested by Simon search is now case-insensitive and makes use of regular expressions.
I tweaked the highlighting as well. For some reason though you need to move the cursor at least once in the text pane before that works correctly...

git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@160 07b9d407-957a-469a-9197-ae794bbef2a3