Issue #48: Delay categorization of decimals
To allow the grammars to handle some decisions which were currently being taken in the tokenizer/lexer stages I performed a big refactoring and added some new grammar features. In brief:

* I reworked the tokenizer stages and added unit tests for each of them.
* I then changed the separator tokenizer to not recognize decimals. Instead those get returned as a sequence of integer, dot (or comma) and integer.
* The tokenizers no longer filter out whitespace or optional separators from the grammar. The grammar can now see everything, if it so desires.
* When the grammar tests for a token it will automatically skip anything it considers whitespace or separators. This gives us back the old, filtering behaviour. There is a new feature, however, which lets the grammar specify when it does not want this filtering: "%noskip". With that it will see all tokens.
* I extended the parser combinators with two new types: "any" (which can match any token), and "tagged", which checks if the next token has a specific tag.

All that allows you to define decimals in the grammar as (approx):

    @INTEGER _ %noskip ( ('.' | ',') @INTEGER _ )
    
(Please check the actual grammar for the full, correct version.

Most everything else are fixes to deal with the new stuff. I removed some old tests which are now redundant or no longer apply. I also removed support for DTD generation, as I don't think anyone is interested in it.

All that now fixes the originally filed problem.

All grammar unit tests still pass, and some new ones were added.

The regression tests still officially fails, but that's only because I haven't updated the expected results yet. In practice parsing of the testsuite looks correct.

Still, this commit is pretty extensive and so stuff may be broken...

git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@165 07b9d407-957a-469a-9197-ae794bbef2a3