Issue: SET statement inside XML PARSE is not working
Issue reported by Suman Saurabh.
It was not really related to the SET statement. The XML PARSE exception handlers were set up to only allow a single statement. Saurabh's example had a SET statement followed by a PERFORM. This triggered the problem.

Updated the grammar to allow multiple statements in the exception handlers.

git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@167 07b9d407-957a-469a-9197-ae794bbef2a3