Sanitized the folder and package structure.
I finally took the time to separate out Koopa more cleanly into its natural parts. These are:
* core (src/core/ and test/core/): this holds the island parser generator which can be used in building island parsers
* cobol (src/cobol/ and test/cobol/): this uses the core to create a Cobol island parser
* apps (src/apps/): combines everything into usable applications (the GUI and the CLI tools)
* examples (src/examples/): has examples showing how to use some Koopa features from the code (e.g. XPath querying)

In theory you should now be able to take just the parser generator (core) from Koopa and use it to build parsers for any language. Which was the main goal behind this cleanup.


git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@172 07b9d407-957a-469a-9197-ae794bbef2a3