Fix for bug found by Marc Dzaebel.
Check LowLevelTest.stage for a unit test of the bug. In short: a dot at the end of a line could combine with a number at the next line to form a decimal, because newlines were always being skipped.

I updated the semantics of newlines to also act as program text. That fixed the bug, but introduced some small regressions with the handling of continuations. The fix for that forms the remainder of this commit.

git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@175 07b9d407-957a-469a-9197-ae794bbef2a3