Start of an alternative implementation of tree parsers.
The existing tree parser support relied entirely on ANTLR to do its thing. But that requires juggling ANTLR tokens files, which is making it harder to add some useful features into Koopa grammars (such as grammar composition, which is next on my most-wanted list).

So the new tree parser support is native to Koopa, without the need for ANTLR or tokens files. It takes regular KG definitions for generating a tree parser. Which means any KG grammar now gets you a matching tree parser "for free".

This commit shows the basic functionality with an accompanying example. The aim is to replace all existing ANTLR tree parsers with this new technology, and then remove the ANTLR tree parser support entirely.



git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@178 07b9d407-957a-469a-9197-ae794bbef2a3