Allowing to set a limit on a part of the grammar.
E.g.:

    %limit sqlStatement %by END-EXEC

This will try parsing an sqlStatement, but only up to the point of an END-EXEC. The sqlStatement rule will never get to see this token.

Applied this to EXEC SQL and EXEC CICS to limit their scope a little. This should help (some) with an issue reported by Suman Saurabh.

git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@187 07b9d407-957a-469a-9197-ae794bbef2a3