Added grammar operator which allows for switch/case like behaviour.
Rather than having a plain choice structure such as:

    ( addStatement
    | displayStatement
    | ...
    )

You can now do:

    $( ADD => addStatement
      | DISPLAY => displayStatement
      | ...
      )

The selector must be a token, and it will not be consumed. I.e. the addStatement will still see the ADD token.
There is also no support for a default/catch-all case.

This new feature helps cut down on the number of tests and backtracking in places where the next token is sufficient for deciding what to do.


git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@193 07b9d407-957a-469a-9197-ae794bbef2a3