Minimized the dependency on ANTLR.
ANTLR's use is now limited to the generation of Koopa's core. We no longer rely on ANTLR's CommonTree in the rest of the code; instead we have now standardized on Koopa's own tree implementation.

Added some extra tests to check XML serialization and XPath evaluation as well, as extra safeguards.

git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@194 07b9d407-957a-469a-9197-ae794bbef2a3