Fixed bug in TreeStream. Tweaked Cobol grammar slightly.
The bug in the TreeStream could be seen when parsing NC109M.CBL from the Cobol85 testsuite. It would show the final paragraph twice in the outline. This was caused due to faulty rollbacking of the tree stream.

The tweaks to the Cobol grammar are just trying to clarify rule for procedureName a bit more.

git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@202 07b9d407-957a-469a-9197-ae794bbef2a3