Issue #51 Grammar issue with COMMIT.
COMMIT is not a reserved word, and can be used as a name.
There is, however, also the use of COMMIT as a verb (found this in the PERCobol Language Reference Manual).
This caused a bad choice in the definition of the PERFORM statement.

I tweaked the grammar rule so it would accepts Inaki GE's example. The solution is not great, but works.

git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@204 07b9d407-957a-469a-9197-ae794bbef2a3