Rework of the CICS grammar.
It no longer provides rules for specific CICS commands, but instead encodes the basic structure of any CICS command.
This means it should be able to accept more inputs.
It also means it may accepts inputs which make no sense (e.g. because a certain option does not belong with a given command).
Given Koopa's rationale, however, this is acceptable.



git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@229 07b9d407-957a-469a-9197-ae794bbef2a3