Allowing grammar rules to be private. And a minor CICS tweak.
Grammar rules can now be marked as private. When you do, that rule becomes private in the generated class. That rule will also no longer appear in the abstract syntax tree.
This allows for building supporting rules without having them add noise to the final parse tree.

I applied this in the CICS grammar to values of options. These can have further nested parentheses, which was not covered previously. The new rule encodes support for balanced parentheses, but will not show up in the parse tree. So you'll still see a single value node in the end.



git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@230 07b9d407-957a-469a-9197-ae794bbef2a3