Support for source format directive. Allow configuration of source and copybook extensions.
The source format directive now gets interpreted by Koopa to do source format switching on the fly. You can still set the initial source format, but from that point on Koopa will toggle the source format based on the directives it finds. Internally the format is being tracked by setting the format as a tag on the tokens, and interpreted as needed by looking it up in the tokens.

In addition, we can now configure extensions for Cobol source files and copybooks separately. The default is equivalent to:
    -Dkoopa.sources=cbl,cob
    -Dkoopa.copybooks=cpy,copy
But you can set this up any way you want. New definitions will replace the default ones completely.

Applying this to the ACAS project, with the default format set to FREE, source extensions 'cbl' and copybook extensions 'cob', only 7 source files still fail to parse.

git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@247 07b9d407-957a-469a-9197-ae794bbef2a3