Auto-keywords
This update adds automatic keyword detection. Where before we had to keep a separate list of possible keywords, and try to keep that up to date and working correctly, now keywords are inferred from the grammar and from the active parsers. Check out AutomaticKeywordsTest for the details of what is going on.

To make this work correctly I did need to allow  literals being treated differently to tokens. Unlike tokens, literals will not be considered as keywords. This helps prevent polluting the active keyword space, as it was doing with intrinsic functions (check Cobol.kg for 'def function').

Related to all this: grammar and unit test fixes to fix regressions. The new method exposed some shortcomings in the existing grammar. Though there are a few places we're doing less well now (esp. conditions vs. arithmetic expressions; that will need some extra love).

Minor KG syntax change: the "not" operator has changed from "-" to "%not". That makes it more visible and readable.

The Breadcrumb was seeing NPEs due to previous changes in Tree's getStart and getEnd. I added in the old versions back as getRawStart and getRawEnd.
Also fixed an issue with the file dialog sometimes not appearing and blocking the app.

git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@258 07b9d407-957a-469a-9197-ae794bbef2a3