Tweaked interaction of separators and markers.
Say you have following statement:

    MOVE A TO B, C

When looking at C, it will be parsed as an identifier. Annoyingly, the comma and space preceding it will also be included in the identifier tree.

This update changes that. Any skipped separators which appear at the start of a rule will get moved outside of the rule.

So, before this update the last identifier would have matched all of ", C". Now, however, it will only match the "C".





git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@271 07b9d407-957a-469a-9197-ae794bbef2a3