UI improvements related to preprocessing
The source view will now show the expanded source code as generated by the preprocessor rather than just the code of the main file.
In addition, navigating the source when the AST view is open will expand the nodes in the AST view.

Other things have been updated to be able to cope with the expanded code view.



git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@292 07b9d407-957a-469a-9197-ae794bbef2a3