New modifier for rules, updated grammer for arithmetic expressions, and tweaks of the grammar for conditions.
While reworking arithmetic expressions to yield cleaner syntax trees I was hitting the limits of efficient backtracking.

The problem was that in order to decide how to structure the syntax tree I had to try some rule first. Do that in too many places and the search explodes to the point where it takes forever to parse anything.

I resolved this by adding a new modifier on grammar rules: 'hiding' (bad name, still looking for something better).
The effect is that a 'hiding' rule will not appear in the syntax tree (so it hides) when it has but a single child. Instead only the single child will be retained.
So for instance:

    hiding rule expr = term (('+' | '-') term)

If there is only one term then the tree will just consist of that term.
If there are two terms with an operator then the tree will be an expr with the terms and operators as children.

I applied this to the arithmetic expressions, and used it to simplify the conditions grammar.



git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@298 07b9d407-957a-469a-9197-ae794bbef2a3