Prevent possible recursion in COPY statement inclusion; and an auto-keyword tweak.
As an example of the first, let's say there exist both a source file "FOO" and a copybook "FOO".
If the sourcefile copies the copybook, by virtue of the lookup mechanism, the COPY statement would resolve to the source file.
This would mean that the source file would include itself. And, of course, it would then hit that COPY again, and repeat the whole pattern.
So now there is a basic test when looking for a copybook that it returns a file other than the one triggering the lookup.

As for the second, when a copybook contained a source file you would expect it to accept the same inputs as regular source files.
However, due to the way auto-keywords worked, a copybook was defining way more keywords than a regular source file. This was causing some valid statements to get rejected.
I resolved this by adding an option which tells a rule (in this case, "copybook") to not track keywords. This breaks the "pollution", and gives the expected behaviour.




git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@314 07b9d407-957a-469a-9197-ae794bbef2a3