Allowing special characters in cobol words in copybooks.
Set koopa.cobolWord.inCopybooks=extended to activate.
Set koopa.cobolWord.extendedCharacters to a string holding the special characters you're allowing. By default this is ":".

Caveat: this only gets applied to separators, for now.

I also extended the GUI to allow setting this, and the max length of cobol words, through a simple dialog.




git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@330 07b9d407-957a-469a-9197-ae794bbef2a3