Allow nested grammar rules; tweaking CALL and PROCEDURE DIVISION header grammar.
Up until now I have been trying to work under the assumption that each concept/name was global and identified a unique thing.
But in order to keep that going I was starting to have to come up with some compound names.

For instance, both the CALL statement and the PROCEDURE DIVISION header define a USING clause and BY REFERENCE, BY VALUE, etc.
But they are all subtly different from each other.
Initially I would just prepend everything with some contextual name, but that's a pretty annoying way forward.

Looking at the spec I found that some of these grammar rules are defined locally (phrased with a "... where ...").
My prepending of a contextual name was basically trying something similar by adding some context or "scope".
So in the end I decided to extend the Koopa DSL to allow local grammar rules.

I applied this to the CALL statement and PROCEDURE DIVISION header (and some other places it made immediate sense).
Roughly speaking, it looks like this:

    def callStatement =
        CALL name [using]
      
      where def using =
          USING ...
      end
    end

So at the end of a grammar rule you can nest more grammar rules simply by prepending it with "where".
The name becomes available within the scope of the top-level grammar rule. 
I.e. in the above example "using" is available anywhere in the "callStatement" or any of its nested rules.

Nested rules can also be accessed from outside, but you have you pass the fully qualified name.
I.e. for "using" in "callStatement" this would be: "callStatement$using". The dollar sign acts as the separator.





git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@342 07b9d407-957a-469a-9197-ae794bbef2a3