REPLACE statements get handled by the preprocessor
This took some refactoring to allow the replacement logic to be shared by the REPLACE and COPY statement.
That refactoring spread into other classes, to help me test things more easily.

This has improvements to the preprocessing grammar as well. Again to allow some sharing between REPLACE and COPY.

And it introduces some more explicit tests for the preprocessing stage in the form of inputs and expected outputs (look for pp_sample files).




git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@345 07b9d407-957a-469a-9197-ae794bbef2a3