KG update to allow quoting keywords, with tests; and SQL grammar extension
The Koopa grammar DSL now allows quoting keywords so that they can be used as names for grammar rules.
A very silly, but legal, example of this would be:

    def `def = `where
      where def `where =  `end  end
    end

I have added grammar unit tests for the Koopa grammar DSL, which were sorely missing.


All of this was to help extend the SQL grammar, where I wanted to use the 'where' keyword as a grammar rule name.

I updated the definitions for table names. 
And also added basic definitions for different cursor-related statements.



git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@351 07b9d407-957a-469a-9197-ae794bbef2a3