Summer cleaning
This is a big update, reworking and standardising the tokenizers, extending the tests and fixing some bugs along the way.

## Tokens

Rather than have each language (re)define its tokenizers we now reuse a common set of tokenizers so that the expected set of tokens is uniform across languages. So, rather than Cobol defining Cobol words as the atom for many things, we now build on words, numbers, strings, etc. instead.

This is helpful for embedded languages (such as CICS and SQL) as these have to share a token stream with the embedding language, which may otherwise interfere.

## Tags

Area and syntactic tags have been cleaned and standardised for use across languages. Cobol specific tags have been moved into the Cobol-handling packages.

## Grammars

Grammars no longer talk about separators, which is a term mostly specific to Cobol. Instead they just identify what part of the source is program text (and will be parsed), and which parts of the program text can be skipped when needed.

Case-sensitivity was also not handled in a standardised way. Now you can ask the Grammar class for a version of any string which may be used in comparison.

## Parser Combinators

Some base classes for implementing parser combinators have been added, providing common functionality.
* GrammaticalCombinator: for combinators which make use of the grammar for skipping skippables.
* UnaryParserDecorator: for combinators which use one other combinator.
* NAryParserDecorator: for combinators which combine many other combinators.

## Testing matches while parsing

A new base class for testing the program text which was matched by another parser combinator has been added, with two incarnations:
* TestForCase: lets you check whether the matched text was all uppercase, lowercase, or not.
* TestForKeyword: lets you check whether the matched text counts as a keyword.

## Sources

Some new classes for sources have been added:
* BashStyleComments: scans the input for bash-style comments and generates tokens marking them as such.
* TagAll: tags all incoming tokens with a certain set of tags.

These are used by the different DSLs.

## Trees

TokenIterator has been removed in favour of TreeWalker.

Support for tree parsers has been completely removed, as the maintenance cost was not offset by its usefulness. There was only one use of it in the codebase itself (CobolOutline), which has been replaced with a solution based on XPath. 

## Grammar tests (.stage)

Grammar tests can now reference each other, allowing definitions for one to be used in another.

There is also a way to use these cross references to generate random test samples and have them be tested. The build script has a target for this (random-cobol-tests) which lets you try it out.

git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@360 07b9d407-957a-469a-9197-ae794bbef2a3