Support for some bottom-up parsing. Big update.
# A syntax for TODOs

When writing grammar rules it is sometimes useful to be able to add a placeholder for something.
Like a todo which you can come back to. This is now possible by using '...'. E.g.:

    def placeholder =  ...  end

When the placeholder gets matched it will just fail. 
So, if it is part of a choice, it will be like it was skipped:

    ( ... | A )

This would just match A.


# Syntax for closures

A new supporting syntax has been added to indicate the creation of closures:

    { something }

This means that `something` will always get matched in the context of its instantiation (its closure).

By itself this would be meaningless, as matching happens at the moment of instantiation, 
but there are a few (new) cases where it does make sense.


# Native code blocks

The syntax for native code blocks has changed, to disambiguate them from closures.
It now looks like:

    =={ /* native code goes here */ }==
    
In the long run, however, I want to remove native code blocks all together.
I only have one use case for them at the moment, which could easily be replaced with a matching XPath query.


# Matching up to a limiter

We now support matching something up to a given limiter. E.g.:

    something %upto { limiter }

When matching `something` it will never see any tokens beyond the limiter.
In addition the overall match will only succeed if `something` matches the tokens
right up to the limiter. 

Note that the limiter must be in a closure. No matter where the matching of `something` has led
the limiter will always get matched in the context of when it was created (here, by the `%upto`).


# Matching with a limiter

Syntax for matching with a limiter has changed. Before it used to be like:

    --- something --> limiter
    
The new syntax is:

    something %before { limiter }

Note that the limiter must be in a closure. No matter where the matching of `something` has led
the limiter will always get matched in the context of when it was created (here, by the `%before`).

This places it in line with the new `%upto` syntax.


# Balancing parentheses, %nested and %notnested

Parsing certain constructs in a lenient way would be much easier if we could account for balancing of parentheses.
E.g. when matching a condition which combines boolean values with `AND` and `OR` it would be great if we could just write:

    def disjunction =  conjunction (OR conjunction)*  end
    def conjunction =  term        (AND term)*        end
    def term        =  _*                             end

Now, the first problem is that `term` would eat all input, including any possible `AND` and `OR`.
This can be resolved using the new `%upto` syntax:

    def disjunction =  conjunction %upto {OR}  (OR  conjunction %upto {OR})*   end
    def conjunction =  term        %upto {AND} (AND term        %upto {AND})*  end
    def term        =  _*                                                      end

Now a term will end at the first mention of an `OR` or an `AND`.
The new problem is that that first mention may be a nested one which shouldn't count:

    ( A OR B ) AND C
    
With the rules as they are this would get parsed as:

    + disjunction
      + term 
        > (
        > A
      + conjunction
        + term
          > B
          > )
        + term
          > C

Which is clearly wrong.

To solve this we introduce three new things: `%balancing`, `%nested` and `notnested`.
Let's rewrite the solution one more time:

    def disjunction =  
        %balancing <{'('} {')'}>
          ( conjunction %upto {%notnested OR}
            ( %notnested OR
              conjunction %upto {%notnested OR}
            )*
          )
    end

    def conjunction =
        term  %upto {%notnested AND}
        ( %notnested AND
          term %upto {%notnested AND}
        )*
    end

    def term =  _*  end

Let's start with the easy part. Note how all mentions of `OR` and `AND` have been preceded by a `%notnested`.
This does what you think. It will only match that `OR` or `AND` when it is not inside parentheses.

How does it know whether it is inside parentheses ? That's what the `%balancing` is for.
The first line in `disjunction` tells the parser to track and pair parentheses while matching.
By itself that is all it does. But when combined with `%notnested` it allows the latter to check the active nesting.

Taken together this means that the parse of the example now becomes:

    + conjunction
      + term
        > (
        > A
        > OR
        > B
        > )
      + term
        > C

Which is correct. The nested disjunction can be handled by letting `term` look test for an opening parenthesis and recurse into `disjunction`.
I won't expand that here, but you can look at `SQL.kg` and `BinaryGrammarTest` to see how it can be done.


# SQL grammar

This now matches the overall structure of search conditions.
It also tries to recognize comparisons and `LIKE` predicates hidden within those conditions.





git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@366 07b9d407-957a-469a-9197-ae794bbef2a3