Optimizer
Added an optimizer on choices and permutations which auto-detects whether they can be optimized through lookahead.
If so the basic choices and permutations get replaced with ones based on dispatchers.

This reduces time spent parsing the Cobol 85 testsuite by a bit more than a third.
Some measurements on my machine:
* Regression test:
** Without optimizer: 65s
** With optimizer: 39s
* Preprocessing regression test:
** Without optimizer: 83s
** With optimizer: 52s

Optimization can be turned off by setting the `koopa.optimize` property to `false`.
I may remove that choice in the long-run, but for now it's a nice fail safe in case there are still some bugs with the optimizer.




git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@387 07b9d407-957a-469a-9197-ae794bbef2a3