Rewrite/Refactoring of the sources.
The intent of the rewrite was to handle source format directives which are embedded in copybooks.
Previously, such directives would change the source format, but that change would be limited to the copybook.
That is, any code following the COPY statement would still be handled as having the original source format.
This was causing trouble when parsing some real industrial code.

This rewrite makes sure that source format directives get applied globally.
That is, a copybook can change the source format in any source file it gets included in.
For an example of this, look at `test/cobol/koopa/cobol/sources/test/All.CopyFree.sample`, which tests this case.

This rewrite restructures the source setups entirely, as can be seen in `src/cobol/koopa/cobol/CobolTokens.java`.
Sources now handle inputs line-by-line, never consuming more lines than needed.
This also allowed the removal of `ThreadedSource`s, which means that all sources are now driven by the parser thread only.

The different `test/cobol/koopa/cobol/sources/test/*.sample` tests have been extended to double check that everything works as expected.
This includes some new abilities to test for the absence of tags as well, rather than just the presence of them.

The `test/cobol/koopa/cobol/parser/preprocessing/test/*.ppsample` tests have also been revised and extended.

And everything has been rerun against the '85 testsuite, as well as against the industrial code which revealed the problem.



git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@388 07b9d407-957a-469a-9197-ae794bbef2a3