Refactoring Source.
At the time it seemed like a good idea to have Source be generic, though I must admit I no longer remember the reasoning.
But these days it has been more of a hindrance than an advantage. 
Having to swith between sources of Data and sources of Token, through the NarrowingSource and WideningSource utility classses was just plain annoying.
So I decided to remove the generics, and simplify the Source interface.

From now on Sources always work on Data, and there no longer is a way to make that more specific.
I have reworked all Source implementations to be able to handle Data when it comes in.
I have updated the Grammar to also reason about Data rather than just Tokens.

Trees are now also Data and it is my intention, in later updates, to pass trees through the sources as well.
Compiler directives, for instance, are parsed into trees, but remain behind in the source which parsed them. If you want to retrieve them you need to look up that source and ask for them.
The idea is that parsed compiler directives will just flow through the sources, and you can capture and process them at the end like any other piece of data.
But that's a step for another commit.




git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@392 07b9d407-957a-469a-9197-ae794bbef2a3