Introducing the CobolProject interface.
Check the docs for the CobolProject interface for details.

In short, the goal is to centralize file specific parsing details, as well as copybook resolution.
The StandardCobolProject duplicates the old behaviour.
But you can define your own CobolProject subclass to tell the CobolParser what to do: which grammar rule to apply, whether to preprocess files, where to find copybooks, etc.

Also moved away from ParsingCoordinator (now CobolParserFactory) in favour of CobolParser. Limited use of the former to the app packages. The cobol packages no longer need it, relying solely on CobolParser and the new CobolProject.

Made a minor tweak to the Cobol grammar, by disallowing empty compilation groups. Empty files can still be accepted if you want to by allowing it in your CobolProject subclass. But the default expects to find something inside a file.



git-svn-id: svn://svn.code.sf.net/p/koopa/code/trunk@397 07b9d407-957a-469a-9197-ae794bbef2a3